+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 1.0.0 (Unreleased)
 
 - **NEW:**  `network-with-security-group` [GH-1]( https://github.com/terraform-alicloud-modules/terraform-alicloud-network-with-security-group/pull/1)
+
+IMPROVEMENTS:
+
+- improve output parameters [GH-2]( https://github.com/terraform-alicloud-modules/terraform-alicloud-network-with-security-group/pull/2)